src/blueguard_agentverse.py
from __future__ import annotations

import asyncio
import os
from datetime import UTC, datetime
from pathlib import Path
from typing import Any
from uuid import uuid4

from dotenv import load_dotenv
from uagents import Agent, Context, Protocol
from uagents_core.contrib.protocols.chat import (
    ChatAcknowledgement,
    ChatMessage,
    EndSessionContent,
    StartSessionContent,
    TextContent,
    chat_protocol_spec,
)

from build_risk_snapshot import (
    build_offline_sample_ships,
    collect_live_ships,
    compute_ship_risk,
    load_whale_sightings,
)
from environmental_context import get_environmental_context

load_dotenv()

AGENT_NAME = os.getenv("BLUEGUARD_AGENT_NAME", "BlueGuard_Navigator").strip() or "BlueGuard_Navigator"
AGENT_SEED = os.getenv("BLUEGUARD_AGENT_SEED", "blueguard navigator default seed replace me").strip()
WHALE_CSV = Path(os.getenv("BLUEGUARD_WHALE_CSV", "data/whale_occurrences_global.csv"))
WHALE_MAX_AGE_DAYS = int(os.getenv("BLUEGUARD_WHALE_MAX_AGE_DAYS", "180"))
ENV_LAT = float(os.getenv("BLUEGUARD_ENV_LAT", "33.73"))
ENV_LON = float(os.getenv("BLUEGUARD_ENV_LON", "-118.26"))
SHIP_LIMIT = int(os.getenv("BLUEGUARD_SHIP_LIMIT", "20"))
SHIP_COLLECTION_SECONDS = int(os.getenv("BLUEGUARD_SHIP_COLLECTION_SECONDS", "15"))
USE_LIVE_AIS = os.getenv("BLUEGUARD_USE_LIVE_AIS", "true").lower() not in {"0", "false", "no"}
AISSTREAM_API_KEY = os.getenv("AISSTREAM_API_KEY", "").strip()
AGENT_ENDPOINT = os.getenv("BLUEGUARD_AGENT_ENDPOINT", "").strip()

agent_kwargs: dict[str, Any] = {"name": AGENT_NAME, "seed": AGENT_SEED}
if AGENT_ENDPOINT:
    agent_kwargs["endpoint"] = [AGENT_ENDPOINT]
agent = Agent(**agent_kwargs)
chat_proto = Protocol(spec=chat_protocol_spec)


def _make_chat_response(text: str, msg_id: Any | None = None) -> ChatMessage:
    return ChatMessage(
        timestamp=datetime.now(UTC),
        msg_id=msg_id or uuid4(),
        content=[TextContent(type="text", text=text)],
    )


async def _collect_ships() -> tuple[list[Any], str]:
    if USE_LIVE_AIS and AISSTREAM_API_KEY:
        try:
            ships = await collect_live_ships(max_ships=SHIP_LIMIT, max_seconds=SHIP_COLLECTION_SECONDS)
            if ships:
                return ships, "live_ais"
        except Exception:
            # Fall through to offline sample ships for reliability during judging.
            pass
    return build_offline_sample_ships(), "offline_sample"


async def _build_blueguard_report() -> dict[str, Any]:
    whales = []
    if WHALE_CSV.exists():
        try:
            whales = load_whale_sightings(path=WHALE_CSV, max_age_days=WHALE_MAX_AGE_DAYS)
        except Exception:
            whales = []

    env_context = get_environmental_context(
        lat=ENV_LAT,
        lon=ENV_LON,
        prefer_live=True,
        timeout_s=20,
    )
    ships, ship_mode = await _collect_ships()
    scored = [compute_ship_risk(ship=ship, whales=whales, env_context=env_context) for ship in ships]
    scored.sort(key=lambda row: row["risk_total"], reverse=True)

    high = [ship for ship in scored if ship["risk_level"] == "HIGH"]
    medium = [ship for ship in scored if ship["risk_level"] == "MEDIUM"]
    top = scored[0] if scored else None
    top_ship_name = top["ship"] if top else "none"
    top_ship_risk = f'{top["risk_level"]} ({top["risk_total"]})' if top else "LOW (0.0)"

    return {
        "generated_at": datetime.now(UTC).isoformat(),
        "ship_mode": ship_mode,
        "ship_count": len(scored),
        "high_count": len(high),
        "medium_count": len(medium),
        "top_ship_name": top_ship_name,
        "top_ship_risk": top_ship_risk,
        "top_ship": top,
        "whale_sightings_used": len(whales),
        "env_zone": env_context.get("high_krill_zone", {}),
    }


def _build_text_reply(user_query: str, report: dict[str, Any]) -> str:
    top_ship = report.get("top_ship") or {}
    action = "Reduce to 10 kts and apply reroute" if top_ship.get("risk_level") == "HIGH" else "Continue monitored transit"
    query_hint = user_query.lower()

    if "premium" in query_hint or "paid" in query_hint:
        return (
            "BlueGuard Premium Safety Report is available. Payment Protocol is optional in this starter and can be enabled next "
            "for paid risk snapshots."
        )

    return (
        f"BlueGuard Navigator status for San Pedro Channel: {report['high_count']} HIGH risk and {report['medium_count']} MEDIUM risk vessels "
        f"out of {report['ship_count']} tracked ships ({report['ship_mode']}). "
        f"Top vessel: {report['top_ship_name']} at {report['top_ship_risk']}. "
        f"Whale context: {report['whale_sightings_used']} sightings from last {WHALE_MAX_AGE_DAYS} days. "
        f"Recommended action: {action}. "
        "This specialist is registered with Chat Protocol for ASI:One discovery."
    )


@chat_proto.on_message(ChatMessage)
async def handle_message(ctx: Context, sender: str, msg: ChatMessage):
    await ctx.send(
        sender,
        ChatAcknowledgement(timestamp=datetime.now(UTC), acknowledged_msg_id=msg.msg_id),
    )

    user_query = ""
    for item in msg.content:
        if isinstance(item, StartSessionContent):
            ctx.logger.info("Session started by %s", sender)
        elif isinstance(item, EndSessionContent):
            ctx.logger.info("Session ended by %s", sender)
        elif isinstance(item, TextContent):
            user_query = item.text or user_query

    report = await _build_blueguard_report()
    reply = _build_text_reply(user_query=user_query, report=report)
    await ctx.send(sender, _make_chat_response(reply, msg.msg_id))


@chat_proto.on_message(ChatAcknowledgement)
async def handle_acknowledgement(ctx: Context, sender: str, msg: ChatAcknowledgement):
    ctx.logger.info(
        "Received acknowledgement from %s for message %s",
        sender,
        msg.acknowledged_msg_id,
    )


agent.include(chat_proto, publish_manifest=True)


if __name__ == "__main__":
    agent.run()
