requirements.txt
websockets>=12.0
uagents>=0.21.0
uagents-core>=0.3.0
python-dotenv>=1.0.1
